refactor(map): fix placement of MapChannel
